requirements.txt
python-telegram-bot[job-queue]==21.6
httpx==0.27.0
python-dotenv==1.0.1
groq==0.9.0
